refactor: move methods to add callbacks to new
class called MutationCallbacks
