+++ packages/graphql_flutter/lib/src/widgets/mutation.dart
@@ -69,80 +69,6 @@ class MutationState extends State<Mutation> {
     }
   }
 
-  OnData get onCompleted {
-    if (widget.onCompleted != null) {
-      return (QueryResult result) {
-        if (!result.loading && !result.optimistic) {
-          return widget.onCompleted(result.data);
-        }
-      };
-    }
-    return null;
-  }
-
-  OnData get onError {
-    if (widget.onError != null) {
-      return (QueryResult result) {
-        if (!result.loading &&
-            result.hasException &&
-            widget.options.errorPolicy != ErrorPolicy.ignore) {
-          return widget.onError(result.exception);
-        }
-      };
-    }
-
-    return null;
-  }
-
-  /// The optimistic cache layer id `update` will write to
-  /// is a "child patch" of the default optimistic patch
-  /// created by the query manager
-  String get _patchId => '${observableQuery.queryId}.update';
-
-  /// apply the user's patch
-  void _optimisticUpdate(QueryResult result) {
-    final Cache cache = client.cache;
-    final String patchId = _patchId;
-    // this is also done in query_manager, but better safe than sorry
-    assert(cache is OptimisticCache,
-        "can't optimisticly update non-optimistic cache");
-    (cache as OptimisticCache).addOptimisiticPatch(patchId, (Cache cache) {
-      widget.update(cache, result);
-      return cache;
-    });
-  }
-
-  // optimistic patches will be cleaned up by the query_manager
-  // cleanup is handled by heirarchical optimism -
-  // as in, because our patch id is prefixed with '${observableQuery.queryId}.',
-  // it will be discarded along with the observableQuery.queryId patch
-  // TODO this results in an implicit coupling with the patch id system
-  OnData get update {
-    if (widget.update != null) {
-      // dereference all variables that might be needed if the widget is disposed
-      final Cache cache = client.cache;
-      final OnMutationUpdate widgetUpdate = widget.update;
-      final OnData optimisticUpdate = _optimisticUpdate;
-
-      // wrap update logic to handle optimism
-      void updateOnData(QueryResult result) {
-        if (result.optimistic) {
-          return optimisticUpdate(result);
-        } else {
-          return widgetUpdate(cache, result);
-        }
-      }
-
-      return updateOnData;
-    }
-    return null;
-  }
-
-  // callbacks will be called against each result in the stream,
-  // which should then rebroadcast queries with the appropriate optimism
-  Iterable<OnData> get callbacks =>
-      <OnData>[onCompleted, update, onError].where(notNull);
-
   /// Run the mutation with the given `variables` and `optimisticResult`,
   /// returning a [MultiSourceResult] for handling both the eager and network results
   MultiSourceResult runMutation(
